refactor(diagnostics): split code catalog modules (#8197)
